Read file, get presigned S3 PUT URL, do HTTP PUT of content
The component is where we should read the file, because we can trigger
the read there, and as part of that event, after reading, call the
HTTP methods as "services".

We first use the API to get a presigned upload URL, with our file's
content type; we also set it to be world readbale, so our NG app can
read from S3 directly for later display.

The services are Observables, and didn't interact well with the reader
when we tried to use it in the putUploadURL service, so we pass in the
file body from the component. We MUST set the content type to match
that of the URL we requested, because NG gratuitously sets the content
type if we do not, and that causes the signature calculation to fail
with no obvious explanation on upload.

Moral of the story: handle the file read and other browsery stuff in
the component, handle API in services with Observables.