Add hander for trigger S3ObjectDeleted, and API DELETE $ID hander
I really like the loose couping: if you (or the API) deletes something
in S3, it fires a trigger to remove it from Dynamo. This way, the API
only has to delete from S3, and Dynamo will track: we don't have ot
have the API delete it from both S3 and Dynamo itself.  This also
means if we MANUALLY delete from S3 (e.g., console) our Dynamo will be accurate.